requirements.txt
openai
pandas
Pillow
tiktoken